#355 * Migrated `core.commands` plugins * Migrated `core.helpandsupport` plugins.
