+++ plugins/core/commands/global.lua
@@ -15,13 +15,16 @@ local commands					= require("cp.commands")
 --------------------------------------------------------------------------------
 -- THE PLUGIN:
 --------------------------------------------------------------------------------
-local plugin = {}
+local plugin = {
+	id				= "core.commands.global",
+	group			= "core",
+}
 
-	--------------------------------------------------------------------------------
-	-- INITIALISE PLUGIN:
-	--------------------------------------------------------------------------------
-	function plugin.init()
-		return commands:new("global"):enable()
-	end
+--------------------------------------------------------------------------------
+-- INITIALISE PLUGIN:
+--------------------------------------------------------------------------------
+function plugin.init()
+	return commands:new("global"):enable()
+end
 
 return plugin
