+++ plugins/core/helpandsupport/credits.lua
@@ -1,50 +1,49 @@
 --------------------------------------------------------------------------------
 --------------------------------------------------------------------------------
---                             F E E D B A C K                                --
+--                             C R E D I T S                                  --
 --------------------------------------------------------------------------------
 --------------------------------------------------------------------------------
 
 --------------------------------------------------------------------------------
 -- EXTENSIONS:
 --------------------------------------------------------------------------------
-local feedback			= require("cp.feedback")
 local metadata			= require("cp.config")
 
 --------------------------------------------------------------------------------
 -- CONSTANTS:
 --------------------------------------------------------------------------------
-local PRIORITY 			= 2
+local PRIORITY 			= 3
 
 --------------------------------------------------------------------------------
 -- THE MODULE:
 --------------------------------------------------------------------------------
 local mod = {}
 
-	function mod.showFeedback()
-		feedback.showFeedback()
+	function mod.openCredits()
+		hs.openAbout()
 	end
 
 --------------------------------------------------------------------------------
 -- THE PLUGIN:
 --------------------------------------------------------------------------------
-local plugin = {}
-
-	--------------------------------------------------------------------------------
-	-- DEPENDENCIES:
-	--------------------------------------------------------------------------------
-	plugin.dependencies = {
-		["cp.plugins.core.menu.helpandsupport"] = "helpandsupport"
+local plugin = {
+	id				= "core.helpandsupport.credits",
+	group			= "core",
+	dependencies	= {
+		["core.menu.helpandsupport"]	= "helpandsupport",
 	}
+}
 
-	--------------------------------------------------------------------------------
-	-- INITIALISE PLUGIN:
-	--------------------------------------------------------------------------------
-	function plugin.init(deps)
-		deps.helpandsupport:addItem(PRIORITY, function()
-			return { title = i18n("provideFeedback"),	fn = mod.showFeedback }
-		end)
+--------------------------------------------------------------------------------
+-- INITIALISE PLUGIN:
+--------------------------------------------------------------------------------
+function plugin.init(deps)
 
-		return mod
-	end
+	deps.helpandsupport:addItem(PRIORITY, function()
+		return { title = i18n("credits"),	fn = mod.openCredits }
+	end)
+	return mod
+
+end
 
 return plugin
